Force re-extraction if dex number is invalid
The dex number is considered invalid if it is not found in
SharedPreferences or the persisted value is less than 1. My only theory
for how this could occur is SharedPreferences corruption.

In this scenario prior to this change, MultiDex would think there were
no secondary dex files to extract and the application would crash at
runtime with ClassNotFoundException.

Bug: 62483082
Change-Id: If3671b53b06bb2fb97a9dd82858a6f43b0e1dc1c
Test: mmma -j frameworks/multidex